+++ src/Sulu/Bundle/ContentBundle/Tests/Functional/Preview/PreviewMessageHandlerTest.php
@@ -43,7 +43,7 @@ class PreviewMessageHandlerTest extends SuluTestCase
 
         $this->component = new PreviewMessageHandler(
             $this->getContainer()->get('sulu_content.preview'),
-            $this->getContainer()->get('sulu_core.webspace.request_analyzer.admin'),
+            $this->getContainer()->get('sulu_core.webspace.request_analyzer'),
             $this->getContainer()->get('doctrine'),
             $this->mapper,
             new NullLogger()