+++ src/Sulu/Bundle/CoreBundle/DependencyInjection/SuluCoreExtension.php
@@ -135,6 +135,7 @@ class SuluCoreExtension extends Extension implements PrependExtensionInterface
         $loader->load('build.xml');
         $loader->load('localization.xml');
         $loader->load('serializer.xml');
+        $loader->load('request_analyzer.xml');
     }
 
     /**